requirements.txt
flask
deepseek-sdk
psycopg2
python-dotenv
slack_bolt